+++ src/main/java/org/folio/circulation/domain/RequestType.java
@@ -1,0 +1,6 @@
+package org.folio.circulation.domain;
+
+public class RequestType {
+  public final static String RECALL = "Recall";
+  public final static String HOLD = "Hold";
+}