+++ NEWS.md
@@ -1,6 +1,7 @@
 ## 5.0.0 Unreleased
 
 * Item status `Checked Out` is now `Checked out` (CIRC-39)
+* Creating a hold or recall request changes the associated item's `status` (CIRC-39)
 * Creating a request for an item that does not exist will now fail (CIRC-39)
 
 ## 4.5.0 2017-10-11